+++ src/main/apis/app/window/windowManager.ts
@@ -9,36 +9,31 @@ class WindowManager implements IWindowManager {
 
   create(name: IWindowList) {
     const windowConfig: IWindowListItem = windowList.get(name)!
-    if (windowConfig.isValid) {
-      if (!windowConfig.multiple) {
-        if (this.has(name)) return this.#windowMap.get(name)!
-      }
-      const window = new BrowserWindow(windowConfig.options())
-      const id = window.id
-      if (windowConfig.multiple) {
-        this.#windowMap.set(`${name}_${window.id}`, window)
-        this.#windowIdMap.set(window.id, `${name}_${window.id}`)
-      } else {
-        this.#windowMap.set(name, window)
-        this.#windowIdMap.set(window.id, name)
-      }
-      windowConfig.callback(window, this)
-      window.on('close', () => {
-        this.deleteById(id)
-      })
-      return window
-    } else {
-      return null
+    if (!windowConfig.isValid) return null
+
+    if (!windowConfig.multiple) {
+      if (this.has(name)) return this.#windowMap.get(name)!
     }
+
+    const window = new BrowserWindow(windowConfig.options())
+    const id = window.id
+    const windowName = windowConfig.multiple ? `${name}_${id}` : name
+
+    this.#windowMap.set(windowName, window)
+    this.#windowIdMap.set(id, windowName)
+
+    windowConfig.callback(window, this)
+    window.on('close', () => {
+      this.deleteById(id)
+    })
+    return window
   }
 
   get(name: IWindowList) {
     if (this.has(name)) {
       return this.#windowMap.get(name)!
-    } else {
-      const window = this.create(name)
-      return window
     }
+    return this.create(name)
   }
 
   has(name: IWindowList) {
@@ -57,11 +52,15 @@ class WindowManager implements IWindowManager {
     const miniWindow = this.#windowMap.get(IWindowList.MINI_WINDOW)
     if (miniWindow && miniWindow.isVisible() && !isSkipMiniWindow) {
       return miniWindow
-    } else {
-      const settingWindow = this.#windowMap.get(IWindowList.SETTING_WINDOW)
-      const trayWindow = this.#windowMap.get(IWindowList.TRAY_WINDOW)
-      return settingWindow || trayWindow || this.create(IWindowList.SETTING_WINDOW)!
     }
+
+    const settingWindow = this.#windowMap.get(IWindowList.SETTING_WINDOW)
+    if (settingWindow) return settingWindow
+
+    const trayWindow = this.#windowMap.get(IWindowList.TRAY_WINDOW)
+    if (trayWindow) return trayWindow
+
+    return this.create(IWindowList.SETTING_WINDOW)!
   }
 }